Changed LoadRom to load a byte array into ROM. Also created a Byte to Hex converter.

+++ D5700ComputerEmulator/src/main/kotlin/instructions/Read.kt
@@ -20,6 +20,6 @@ class Read(
     override fun perform() {
         require(mySplit[0] == 3u) { "First 4 bits of first byte must be 3 for Read Instruction." }
         cpu.registers[mySplit[1].toInt()] =
-            memory[cpu.memory.toInt()].read(cpu.address.toInt()).toUByte()
+            memory[cpu.memory.toInt()].read(cpu.address.toInt())
     }
 }
